+++ src/gbz/tests.rs
@@ -235,6 +235,14 @@ fn serialize() {
 }
 
 #[test]
+fn serialize_versions() {
+    let filename = support::get_test_data("example.gbz");
+    let gbz: GBZ = serialize::load_from(&filename).unwrap();
+    let lossless_versions: Vec<usize> = (GBZ::MIN_VERSION..=GBZ::MAX_VERSION).collect();
+    serialize::test_versions(&gbz, "gbz", &lossless_versions);
+}
+
+#[test]
 fn nodes() {
     let filename = support::get_test_data("example.gbz");
     let gbz: GBZ = serialize::load_from(&filename).unwrap();
@@ -327,6 +335,14 @@ fn serialize_trans() {
     let filename = support::get_test_data("translation.gbz");
     let gbz: GBZ = serialize::load_from(&filename).unwrap();
     serialize::test(&gbz, "gbz-trans", None, true);
+}
+
+#[test]
+fn serialize_trans_versions() {
+    let filename = support::get_test_data("translation.gbz");
+    let gbz: GBZ = serialize::load_from(&filename).unwrap();
+    let lossless_versions: Vec<usize> = (GBZ::MIN_VERSION..=GBZ::MAX_VERSION).collect();
+    serialize::test_versions(&gbz, "gbz-trans", &lossless_versions);
 }
 
 #[test]
@@ -569,44 +585,48 @@ fn reference_positions() {
 //-----------------------------------------------------------------------------
 
 #[test]
-fn v1_files() {
+fn gbz_versions() {
     let graphs = vec![
         (support::get_test_data("example-v1.gbz"), support::get_test_data("example.gbz")),
+        (support::get_test_data("example-v2.gbz"), support::get_test_data("example.gbz")),
+        (support::get_test_data("example-v3.gbz"), support::get_test_data("example.gbz")),
         (support::get_test_data("translation-v1.gbz"), support::get_test_data("translation.gbz")),
+        (support::get_test_data("translation-v2.gbz"), support::get_test_data("translation.gbz")),
+        (support::get_test_data("translation-v3.gbz"), support::get_test_data("translation.gbz")),
     ];
 
-    for (v1_file, v2_file) in graphs.iter() {
-        let v1_graph = serialize::load_from::<GBZ, _>(&v1_file);
-        assert!(v1_graph.is_ok(), "Could not load v1 graph from file {}", v1_file.to_string_lossy());
-        let v1_graph = v1_graph.unwrap();
+    for (versioned_file, reference_file) in graphs.iter() {
+        let versioned_graph = serialize::load_from::<GBZ, _>(&versioned_file);
+        assert!(versioned_graph.is_ok(), "Could not load versioned graph from file {}", versioned_file.to_string_lossy());
+        let versioned_graph = versioned_graph.unwrap();
 
-        let v2_graph = serialize::load_from::<GBZ, _>(&v2_file);
-        assert!(v2_graph.is_ok(), "Could not load v2 graph from file {}", v2_file.to_string_lossy());
-        let v2_graph = v2_graph.unwrap();
+        let reference_graph = serialize::load_from::<GBZ, _>(&reference_file);
+        assert!(reference_graph.is_ok(), "Could not load reference graph from file {}", reference_file.to_string_lossy());
+        let reference_graph = reference_graph.unwrap();
 
-        assert_eq!(v1_graph.nodes(), v2_graph.nodes(), "Invalid number of nodes in v1 graph loaded from file {}", v1_file.to_string_lossy());
-        for node_id in v1_graph.node_iter() {
-            assert_eq!(v1_graph.sequence(node_id), v2_graph.sequence(node_id), "Invalid sequence for node {} in v1 graph loaded from file {}", node_id, v1_file.to_string_lossy());
-            let v1_succ = v1_graph.successors(node_id, Orientation::Forward).unwrap();
-            let v2_succ = v2_graph.successors(node_id, Orientation::Forward).unwrap();
-            assert!(v1_succ.eq(v2_succ), "Invalid successors for node {} in v1 graph loaded from file {}", node_id, v1_file.to_string_lossy());
-            let v1_pred = v1_graph.predecessors(node_id, Orientation::Forward).unwrap();
-            let v2_pred = v2_graph.predecessors(node_id, Orientation::Forward).unwrap();
-            assert!(v1_pred.eq(v2_pred), "Invalid predecessors for node {} in v1 graph loaded from file {}", node_id, v1_file.to_string_lossy());
+        assert_eq!(versioned_graph.nodes(), reference_graph.nodes(), "Invalid number of nodes in versioned graph loaded from file {}", versioned_file.to_string_lossy());
+        for node_id in versioned_graph.node_iter() {
+            assert_eq!(versioned_graph.sequence(node_id), reference_graph.sequence(node_id), "Invalid sequence for node {} in versioned graph loaded from file {}", node_id, versioned_file.to_string_lossy());
+            let versioned_succ = versioned_graph.successors(node_id, Orientation::Forward).unwrap();
+            let reference_succ = reference_graph.successors(node_id, Orientation::Forward).unwrap();
+            assert!(versioned_succ.eq(reference_succ), "Invalid successors for node {} in versioned graph loaded from file {}", node_id, versioned_file.to_string_lossy());
+            let versioned_pred = versioned_graph.predecessors(node_id, Orientation::Forward).unwrap();
+            let reference_pred = reference_graph.predecessors(node_id, Orientation::Forward).unwrap();
+            assert!(versioned_pred.eq(reference_pred), "Invalid predecessors for node {} in versioned graph loaded from file {}", node_id, versioned_file.to_string_lossy());
         }
 
-        assert_eq!(v1_graph.has_translation(), v2_graph.has_translation(), "Invalid translation presence in v1 graph loaded from file {}", v1_file.to_string_lossy());
-        if v1_graph.has_translation() {
-            let v1_segments = v1_graph.segment_iter().unwrap();
-            let v2_segments = v2_graph.segment_iter().unwrap();
-            assert!(v1_segments.eq(v2_segments), "Invalid segments in v1 graph loaded from file {}", v1_file.to_string_lossy());
+        assert_eq!(versioned_graph.has_translation(), reference_graph.has_translation(), "Invalid translation presence in versioned graph loaded from file {}", versioned_file.to_string_lossy());
+        if versioned_graph.has_translation() {
+            let versioned_segments = versioned_graph.segment_iter().unwrap();
+            let reference_segments = reference_graph.segment_iter().unwrap();
+            assert!(versioned_segments.eq(reference_segments), "Invalid segments in versioned graph loaded from file {}", versioned_file.to_string_lossy());
         }
 
-        assert_eq!(v1_graph.paths(), v2_graph.paths(), "Invalid number of paths in v1 graph loaded from file {}", v1_file.to_string_lossy());
-        for path_id in 0..v1_graph.paths() {
-            let v1_path = v1_graph.path(path_id, Orientation::Forward).unwrap();
-            let v2_path = v2_graph.path(path_id, Orientation::Forward).unwrap();
-            assert!(v1_path.eq(v2_path), "Invalid path {} in v1 graph loaded from file {}", path_id, v1_file.to_string_lossy());
+        assert_eq!(versioned_graph.paths(), reference_graph.paths(), "Invalid number of paths in versioned graph loaded from file {}", versioned_file.to_string_lossy());
+        for path_id in 0..versioned_graph.paths() {
+            let versioned_path = versioned_graph.path(path_id, Orientation::Forward).unwrap();
+            let reference_path = reference_graph.path(path_id, Orientation::Forward).unwrap();
+            assert!(versioned_path.eq(reference_path), "Invalid path {} in versioned graph loaded from file {}", path_id, versioned_file.to_string_lossy());
         }
     }
 }